sav-163: display full answers for autocomplete questions
